# openapi.yaml
openapi: 3.0.0
info:
  title: Mlog API
  description: API Gateway for post Microservices (Content, Engagement)
  version: 1.0.0
paths:
...
  /api/v1/users:
    get:
      summary: Get all users
      operationId: listAllUsers
      tags: [users]
      responses: 
        "200":
          description: get all users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/userResponse"
    post:
      summary: Create user
      operationId: createUser
      tags: [users]
      parameters:
        - name: X-Userinfo
          in: header
          required: true
          description: >-
            A base64-encoded JSON object containing user claims forwarded from the API gateway.
            After decoding, the JSON will contain user information such as subject (user ID), roles, etc.
            Example decoded object: `{"sub": "user-uuid", "realm_access": {"roles": ["user"]}, "email": "user@example.com"}`
          schema:
            type: string
            format: base64
            example: eyJzdWIiOiAiMTIzNDU2Nzg5MCIsICJuYW1lIjogIkpvaG4gRG9lIiwgImlhdCI6IDE1MTYyMzkwMjJ9
      requestBody:
        description: body of user create request
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/userRequest"
      responses:
        "201":
          description: created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/userResponse"
  /api/v1/users/{user_id}:
    get:
      summary: Get a specific user with the user_id
      operationId: getUser 
      tags: [users]
      parameters:
        - name: user_id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses: 
        "200":
          description: Get a specific user with the user_id
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/userResponse"
  /api/v1/users/me:
    get:
      summary: Get my user info
      operationId: getCurrentUser
      tags: [users]
      parameters:
        - name: X-Userinfo
          in: header
          required: true
          description: >-
            A base64-encoded JSON object containing user claims forwarded from the API gateway.
            After decoding, the JSON will contain user information such as subject (user ID), roles, etc.
            Example decoded object: `{"sub": "user-uuid", "realm_access": {"roles": ["user"]}, "email": "user@example.com"}`
          schema:
            type: string
            format: base64
            example: eyJzdWIiOiAiMTIzNDU2Nzg5MCIsICJuYW1lIjogIkpvaG4gRG9lIiwgImlhdCI6IDE1MTYyMzkwMjJ9
      responses: 
        "200":
          description: Get my user info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/userResponse"
    put:
      summary: Edit my info
      operationId: updateCurrentUser
      tags: [users]
      parameters:
        - name: X-Userinfo
          in: header
          required: true
          description: >-
            A base64-encoded JSON object containing user claims forwarded from the API gateway.
            After decoding, the JSON will contain user information such as subject (user ID), roles, etc.
            Example decoded object: `{"sub": "user-uuid", "realm_access": {"roles": ["user"]}, "email": "user@example.com"}`
          schema:
            type: string
            format: base64
            example: eyJzdWIiOiAiMTIzNDU2Nzg5MCIsICJuYW1lIjogIkpvaG4gRG9lIiwgImlhdCI6IDE1MTYyMzkwMjJ9
    requestBody:
      required: true
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/UpdateUserRequest"
      responses: 
      "200":
        description: User updated successfully
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserResponse"
      "400":
        description: Invalid request
      "401":
        description: Unauthorized


  # some basic admin routes
...
components:
  schemas:
    userRequest:
      type: object
      properties:
        username:
          type: string
          description: username, must be unique
          example: 6myduck10293
        display_name:
          type: string
          description: user's display name
          example: 6myduck
        bio:
          type: string
          description: user's bio
          example: I'm the lord of all ducks
        avatar_url: 
          type: string
          description: Profile picture for the post.
          example: https://ipfs.io/ipfs/QmQA66xJFVcP5h8fs1PJt3fXeGwVPB55Q6o8PwEx1Jc2kz
      required:
        - username
...
    userResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: user's id, autoincrement
          example: 1
        username:
          type: string
          description: username, must be unique
          example: 6myduck10293
        display_name:
          type: string
          description: user's display name
          example: 6myduck
        bio:
          type: string
          description: user's bio
          example: I'm the lord of all ducks
        avatar_url: 
          type: string
          description: Profile picture for the post.
          example: https://ipfs.io/ipfs/QmQA66xJFVcP5h8fs1PJt3fXeGwVPB55Q6o8PwEx1Jc2kz
        banned:
          type: boolean
          description: Whether the user is banned
        created_at:
          type: string
          format: date-time
          description: When the user created their account
          example: date-time
        updated_at:
          type: string
          format: date-time
          description: When the user info was last modified
          example: date-time

    # --- POST SCHEMAS ---
...