# services/user/src/main/java/com/mlog/user/api/UsersApi.java
/*
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.17.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package com.mlog.user.api;

import java.util.UUID;
import com.mlog.user.model.UserRequest;
import com.mlog.user.model.UserResponse;
import io.swagger.v3.oas.annotations.ExternalDocumentation;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import io.swagger.v3.oas.annotations.media.ExampleObject;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.multipart.MultipartFile;

import jakarta.validation.Valid;
import jakarta.validation.constraints.*;
import java.util.List;
import java.util.Map;
import jakarta.annotation.Generated;

@Validated
@Tag(name = "users", description = "the users API")
public interface UsersApi {

    default UsersApiDelegate getDelegate() {
        return new UsersApiDelegate() {};
    }

    String PATH_CREATE_USER = "/api/v1/users";
    /**
     * POST /api/v1/users : Create user
     *
     * @param xUserinfo A base64-encoded JSON object containing user claims forwarded from the API gateway. After decoding, the JSON will contain user information such as subject (user ID), roles, etc. Example decoded object: &#x60;{\&quot;sub\&quot;: \&quot;user-uuid\&quot;, \&quot;realm_access\&quot;: {\&quot;roles\&quot;: [\&quot;user\&quot;]}, \&quot;email\&quot;: \&quot;user@example.com\&quot;}&#x60; (required)
     * @param userRequest body of user create request (required)
     * @return created (status code 201)
     */
    @Operation(
        operationId = "createUser",
        summary = "Create user",
        tags = { "users" },
        responses = {
            @ApiResponse(responseCode = "201", description = "created", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = UserResponse.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = UsersApi.PATH_CREATE_USER,
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    default ResponseEntity<UserResponse> createUser(
        @NotNull @Parameter(name = "X-Userinfo", description = "A base64-encoded JSON object containing user claims forwarded from the API gateway. After decoding, the JSON will contain user information such as subject (user ID), roles, etc. Example decoded object: `{\"sub\": \"user-uuid\", \"realm_access\": {\"roles\": [\"user\"]}, \"email\": \"user@example.com\"}`", required = true, in = ParameterIn.HEADER) @RequestHeader(value = "X-Userinfo", required = true) String xUserinfo,
        @Parameter(name = "UserRequest", description = "body of user create request", required = true) @Valid @RequestBody UserRequest userRequest
    ) {
        return getDelegate().createUser(xUserinfo, userRequest);
    }


    String PATH_GET_CURRENT_USER = "/api/v1/users/me";
    /**
     * GET /api/v1/users/me : Get my user info
     *
     * @param xUserinfo A base64-encoded JSON object containing user claims forwarded from the API gateway. After decoding, the JSON will contain user information such as subject (user ID), roles, etc. Example decoded object: &#x60;{\&quot;sub\&quot;: \&quot;user-uuid\&quot;, \&quot;realm_access\&quot;: {\&quot;roles\&quot;: [\&quot;user\&quot;]}, \&quot;email\&quot;: \&quot;user@example.com\&quot;}&#x60; (required)
     * @return Get my user info (status code 200)
     */
    @Operation(
        operationId = "getCurrentUser",
        summary = "Get my user info",
        tags = { "users" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Get my user info", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = UserResponse.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = UsersApi.PATH_GET_CURRENT_USER,
        produces = { "application/json" }
    )
    default ResponseEntity<UserResponse> getCurrentUser(
        @NotNull @Parameter(name = "X-Userinfo", description = "A base64-encoded JSON object containing user claims forwarded from the API gateway. After decoding, the JSON will contain user information such as subject (user ID), roles, etc. Example decoded object: `{\"sub\": \"user-uuid\", \"realm_access\": {\"roles\": [\"user\"]}, \"email\": \"user@example.com\"}`", required = true, in = ParameterIn.HEADER) @RequestHeader(value = "X-Userinfo", required = true) String xUserinfo
    ) {
        return getDelegate().getCurrentUser(xUserinfo);
    }


    String PATH_GET_USER = "/api/v1/users/{user_id}";
    /**
     * GET /api/v1/users/{user_id} : Get a specific user with the user_id
     *
     * @param userId  (required)
     * @return Get a specific user with the user_id (status code 200)
     */
    @Operation(
        operationId = "getUser",
        summary = "Get a specific user with the user_id",
        tags = { "users" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Get a specific user with the user_id", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = UserResponse.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = UsersApi.PATH_GET_USER,
        produces = { "application/json" }
    )
    default ResponseEntity<UserResponse> getUser(
        @NotNull @Parameter(name = "user_id", description = "", required = true, in = ParameterIn.PATH) @PathVariable("user_id") UUID userId
    ) {
        return getDelegate().getUser(userId);
    }


    String PATH_LIST_ALL_USERS = "/api/v1/users";
    /**
     * GET /api/v1/users : Get all users
     *
     * @return get all users (status code 200)
     */
    @Operation(
        operationId = "listAllUsers",
        summary = "Get all users",
        tags = { "users" },
        responses = {
            @ApiResponse(responseCode = "200", description = "get all users", content = {
                @Content(mediaType = "application/json", array = @ArraySchema(schema = @Schema(implementation = UserResponse.class)))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = UsersApi.PATH_LIST_ALL_USERS,
        produces = { "application/json" }
    )
    default ResponseEntity<List<UserResponse>> listAllUsers(
        
    ) {
        return getDelegate().listAllUsers();
    }

}